+++ NEWS.md
@@ -1,5 +1,9 @@
 # printify 1.0.1 - DEVELOPMENT
 
+### New functionality
+
+* `set_no_color()`: Suppresses the color codes so that messages can be printed clean. (23.04.2026)
+
 ### Additionally
 
 * Message time stamps are now drawn in a lighter grey to enhance visibility on darker themes. (20.04.2026)